requirements.txt
Flask
Flask-Login
Flask-SocketIO
Flask-SQLAlchemy
paho-mqtt
python-dotenv
eventlet
opencv-python
